+++ src/mvc/servlet/GoUpdateReportServlet.java
@@ -13,44 +13,51 @@ import mvc.vo.*;
 public class GoUpdateReportServlet extends HttpServlet {
 	public void doGet(HttpServletRequest req, HttpServletResponse resp)
 			throws ServletException, IOException {
-		//set varible
-		HttpSession session = req.getSession();
-		String userid = session.getAttribute("userid").toString();	
-		String prjPath = req.getSession().getServletContext().getRealPath("/");
-		String path = "editreport.jsp";
-		int reportid = Integer.parseInt(req.getParameter("reportid"));	
-		ArrayList<Attach> attachlist = new ArrayList<Attach>();
-		//get report
-		Report report = new Report();
-		report.setreportid(reportid);
-		report.setuserid(userid);
-		try {
-			DAOFactory.getIReportDAOInstance().readReportById(report);
-		} catch (Exception e) {
-			e.printStackTrace();
+		//session verify
+		if(req.getSession(false).getAttribute("userid")==null){
+			String errorpath = "sessionloss.jsp";
+			req.getRequestDispatcher(errorpath).forward(req, resp);
 		}
-		//delete the useless attachment	
-		Attach attach_del = new Attach();
-		attach_del.setUserId(userid);
-		try {
-			DAOFactory.getIAttachDAOInstance().deleteEmptyReportId(attach_del,prjPath);
-		} catch (Exception e) {
-			e.printStackTrace();
+		else{
+			//set varible
+			HttpSession session = req.getSession();
+			String userid = session.getAttribute("userid").toString();	
+			String prjPath = req.getSession().getServletContext().getRealPath("/");
+			String path = "editreport.jsp";
+			int reportid = Integer.parseInt(req.getParameter("reportid"));	
+			ArrayList<Attach> attachlist = new ArrayList<Attach>();
+			//get report
+			Report report = new Report();
+			report.setreportid(reportid);
+			report.setuserid(userid);
+			try {
+				DAOFactory.getIReportDAOInstance().readReportById(report);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			//delete the useless attachment	
+			Attach attach_del = new Attach();
+			attach_del.setUserId(userid);
+			try {
+				DAOFactory.getIAttachDAOInstance().deleteEmptyReportId(attach_del,prjPath);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			////get attachment of report
+			Attach attach = new Attach();
+			attach.setReportid(reportid);
+			attach.setUserId(userid);
+			try {
+				attachlist = DAOFactory.getIAttachDAOInstance().readAttachByReportId(attach);
+				//System.out.println("ViewReportServlet-attachlist-reportid 0:"+attachlist.get(0).getReportid());
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			//jump to page
+			req.setAttribute("report", report);
+			req.setAttribute("attachlist", attachlist);
+			req.getRequestDispatcher(path).forward(req, resp);
 		}
-		////get attachment of report
-		Attach attach = new Attach();
-		attach.setReportid(reportid);
-		attach.setUserId(userid);
-		try {
-			attachlist = DAOFactory.getIAttachDAOInstance().readAttachByReportId(attach);
-			//System.out.println("ViewReportServlet-attachlist-reportid 0:"+attachlist.get(0).getReportid());
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		//jump to page
-		req.setAttribute("report", report);
-		req.setAttribute("attachlist", attachlist);
-		req.getRequestDispatcher(path).forward(req, resp);
 	}
 
 	public void doPost(HttpServletRequest req, HttpServletResponse resp)
